+++ lib/collaboration_web/controllers/download_controller.ex
@@ -1,0 +1,7 @@
+defmodule CollaborationWeb.DownloadController do
+  use CollaborationWeb, :controller
+
+  def index(conn, _) do
+    render conn, "index.html"
+  end
+end